# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: POWAPLAY API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List public projects for discovery
      parameters:
        - name: tag
          in: query
          required: false
          schema:
            type: string
        - name: style
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of public projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/{username}/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Get a user's public profile
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /users/{username}/projects:
    get:
      operationId: getUserProjects
      tags: [users]
      summary: List a user's public projects
      parameters:
        - name: username
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User's public projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current authenticated user info
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUserEnvelope"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /me/projects:
    get:
      operationId: getMyProjects
      tags: [projects]
      summary: List authenticated user's projects (including hidden)
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: User's projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /me/projects/{id}:
    patch:
      operationId: updateMyProject
      tags: [projects]
      summary: Update a project's visibility or metadata
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
...
  /me/notifications:
    get:
      operationId: getMyNotifications
      tags: [notifications]
      summary: Get user's notifications
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /me/notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ProjectDetail:
      type: object
      required: [id, title, slug, url, description, tags, style, isPublic, isHidden, thumbnailUrl, previewVideoUrl, favoriteCount, ownerUsername, ownerAvatarUrl, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        url:
          type: string
        description:
          type: [string, 'null']
        tags:
          type: array
          items:
            type: string
        style:
          type: [string, 'null']
        isPublic:
          type: boolean
        isHidden:
          type: boolean
        thumbnailUrl:
          type: [string, 'null']
        previewVideoUrl:
          type: [string, 'null']
        favoriteCount:
          type: integer
        ownerUsername:
          type: string
        ownerAvatarUrl:
          type: [string, 'null']
        createdAt:
          type: string
          format: date-time
    ProjectListResponse:
      type: object
      required: [projects, total]
      properties:
        projects:
          type: array
          items:
            $ref: "#/components/schemas/ProjectDetail"
        total:
          type: integer
    UpdateProjectBody:
      type: object
      properties:
        isHidden:
          type: boolean
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        style:
          type: string
...
    UserProfile:
      type: object
      required: [id, username, displayName, avatarUrl, projectCount, createdAt]
      properties:
        id:
          type: string
        username:
          type: string
        displayName:
          type: [string, 'null']
        avatarUrl:
          type: [string, 'null']
        bio:
          type: [string, 'null']
        projectCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    NotificationDetail:
      type: object
      required: [id, type, title, read, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
        title:
          type: string
        body:
          type: [string, 'null']
        feedbackId:
          type: [integer, 'null']
        actionUrl:
          type: [string, 'null']
        read:
          type: boolean
        createdAt:
          type: string
          format: date-time
    NotificationListResponse:
      type: object
      required: [notifications, unreadCount]
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/NotificationDetail"
        unreadCount:
          type: integer